Refactor extract status.Checker and Service interfaces
This moves a lot of code around, but the code decision are:

- extracting status.Checker and status.Service as interfaces
- implementing them in the securedrop package

Checking a service status requires an HTTP client.
Each service defined how they expose their status information
and how it shold be requested.
A SecureDrop instance is a service which status can be checked.

As far as the status.checker is concerned, there is no concurrency.

It would be good to be able to ensure a service doesn't hold on
the checker for too long, but that is left for later.

It seems resonable to me for a securedrop.Instance to store
its metadata, and be updated on status checks (that was already
the case before I renamed sdstatus.Info into securedrop.Instance).
Doing that required acknowledging that the service is specifically
a securedrop.Instance, note the type assertion.

For some reason, I'm not sure about the instance Available field...
so for now I left it out ot the Check function.

Table-driven tests are convenient!
Error wrapping (%w verb) is also a feature of Go that I like.

The status.Checker encapsulate the HTTP client side-effects in
a private field that can be overriden when testing (same package).