Refactor extract the CLi app constructor
The local App type embeds the urfave/cli type and by doing so
exposes all the same fields in addition to its own.

The interest of defining a local App yype here is to allow the scan
function to be defined as a field that can be overriden when testing.

That should allow to conveniently test the handling of errors without
needing to deal with the side-effects of the scanning, but I'm not sure
if this would be sufficient yet.